Added support for files in @covers annotations.
If the value after the @covers annotation contains a period (.),
it is assumed to be a file. Absolute and relative paths are acceptable,
and the include path is searched.

Every line of the file will be available for code coverage tracking.
This is most useful for Zend view scripts (.phtml) and other files
that contain PHP code without a class definition.