Fix for HHVM code coverage to bring the it close to real #302
fb_get_code_coverage returns the number of time that a line in a file
was executed which is totaly different what standard xdebug
implementation is. This patch bring the number of executed lines close
to what is supposed to be. However still needs more work on some edge
cases.