Add custom classes for progress bars
As content security policies are supported across more and more
browsers, the inline styling used for the progress bar widths won't be
processed. Instead of requiring everyone to add a bunch of boring CSP
rules to support this, we add some custom classes for each of the
possible progress bar widths.

We assume that a content security policy allows loading of the
stylesheets, since if it doesn't then there's nothing we can do
anyway.

Fixes #538