Fix PHP Error on mkdir() on existing directory
I cannot do code-coverage because we have a custom error handler that converts PHP Errors into exceptions.

It is bad logic to intentionally yield errors. The problem here was that mkdir() was executed before is_dir(). I reversed the direction of these calls (is_dir() before mkdir()) and made them into a function, since the same logic was three times in the class.