Implement SebastianBergmann\CodeCoverage\CodeCoverage::forceStaticAnalysisCache() and SebastianBergmann\CodeCoverage\CodeCoverage::doNotForceStaticAnalysisCache()
The functions file_exists() and filemtime() are used by
SebastianBergmann\CodeCoverage\StaticAnalysis\Cache::cacheHas() to
determine whether a static analysis cache file exists and is up-to-date.
These functions are slow in some environments. So slow, in fact, that not
using the static analysis cache can be faster than using it.

This commit introduces an API that makes it possible to disable the
cache validation, thus avoiding the usage of the aforementioned
functions.

Note: Validation using file_exists() and filemtime() for the static
analysis cache must not be disabled when there is no guarantee that all
relevant cache files exists and are up-to-date.