Add interface to specify custom.css file for HTML report
Related to #556.

It would help to make the custom CSS feature more accessible by making
the file a configuration option, e.g. in phpunit:

    <coverage>
        <report>
            <html customCssFile="path/to/custom.css" />
        </report>
    </coverage>

The `$customCssFile` argument is added to the Html\Facade interface as
a prerequisite to any changes to phpunit configuration options. If not
null, it copies this file to the custom.css destination instead of the
empty custom.css stub file.