Use Version::id() instead of "MD5 over source code" for encoding static analysis cache format in cache keys for static analysis cache
This means that cached static analysis results will be treated as
outdated when a new/different version of php-code-coverage is used even
if the static analysis code has not changed. Optimizing this edge case
for performance is not worth risking errors in the code for calculating
the "MD5 over source code" that we previously used for this.